pyproject.toml
[project]
name = "sideral"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.88.0",
    "huggingface-hub>=1.8.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openai>=2.30.0",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "streamlit>=1.55.0",
    "tensorflow>=2.21.0",
    "tensorflow-cpu>=2.21.0",
]

[tool.uv.workspace]
members = [
    "sideral",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "ruff>=0.15.9",
]
